machine-learning/backend-android/requirements.txt
Flask
gunicorn
google-cloud-firestore
python-dotenv
google-auth
requests